[FEAT] 6주차 미션_맹덕

+++ mengduck/week4/umc10th-yongjae/src/main/java/org/example/umc10thyongjae/global/apiPayload/handler/HeaderAuthInterceptor.java
@@ -10,7 +10,7 @@ public class HeaderAuthInterceptor implements HandlerInterceptor {
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
         String token = request.getHeader("Authorization");
 
-        int userKey = 1234;
+        long userKey = 1234;
 
         request.setAttribute("userKey", userKey);
